html_frontend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv 